Fix TypeScript imports for hooks feature - resolved circular dependencies
